+++ frontend/src/hooks/context/index.ts
@@ -1,1 +1,3 @@
 export { default as useCategoryContext } from './useCategoryContext';
+export { default as useMemberValueContext } from './useMemberValueContext';
+export { default as useMemberActionContext } from './useMemberActionContext';
